+++ actix-connect/CHANGES.md
@@ -1,5 +1,11 @@
 # Changes
 
+## [1.0.0-alpha.3] - 2019-12-xx
+
+### Changed
+
+* Migrate to `tokio=0.2.2`
+
 ## [1.0.0-alpha.2] - 2019-12-02
 
 ### Changed