+++ actix-server/CHANGES.md
@@ -2,6 +2,10 @@
 
 ## [1.0.0-alpha.3] - 2019-12-xx
 
+### Changed
+
+* Migrate to `tokio=0.2.2`
+
 ### Fixed
 
 * Fix compilation on non-unix platforms