+++ actix-utils/CHANGES.md
@@ -2,6 +2,8 @@
 
 ## [1.0.0-alpha.3] - 2019-12-xx
 
+* Migrate to `tokio=0.2.2`
+
 * Fix oneshot
 
 ## [1.0.0-alpha.2] - 2019-12-02